pyproject.toml
[tool.poetry]
name = "whostweetingwho"
version = "0.1.0"
description = ""
authors = ["Jiahe <yi.jiahe00@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "\b"
python-dotenv = "^0.19.2"
elasticsearch = "^7.16.1"
pytest = "^6.2.5"
responses = "^0.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
